#33: renamed packages based on groupId and artifactId.
