+++ build.gradle
@@ -12,7 +12,7 @@ version '2.0.2'
 description 'A library and utility to compare different images.'
 sourceCompatibility = 1.8
 
-mainClassName = "ua.comparison.image.ImageComparison"
+mainClassName = "com.github.romankh3.image.comparison.ImageComparison"
 
 repositories {
     mavenCentral()
