+++ src/test/java/com/github/romankh3/image/comparison/ArgsParserUnitTest.java
@@ -1,4 +1,4 @@
-package ua.comparison.image;
+package com.github.romankh3.image.comparison;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
